did some changes
