+++ rh-tictactoe/src/app/screens/Login/components/customInput/styles.scss
