--- rh-tictactoe/src/constants/formNames.js
@@ -1,1 +1,0 @@
-export const Login = 'login';
