+++ rh-tictactoe/src/index.js
@@ -3,7 +3,6 @@ import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 
 import { store } from './redux/store';
-// import Game from './app/screens/Game';
 import Login from './app/screens/Login';
 import './scss/index.scss';
